Actually generate the access token, and return the correct expiry period

+++ service/src/main/kotlin/uk/co/grahamcox/goworlds/service/oauth2/http/AccessTokenModel.kt
@@ -13,7 +13,7 @@ data class AccessTokenModel(
         val tokenType: String,
 
         @JsonProperty("expires_in")
-        val expiry: Int,
+        val expiry: Long,
 
         @JsonProperty("scope")
         val scope: String? = null,
